refactor(ts): google find uses core/generic walkFind; statEntry swallows only ENOENT
