--- typescript/packages/core/src/core/google/find.ts
@@ -1,154 +1,0 @@
-// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
-
-import type { IndexCacheStore } from '../../cache/index/store.ts'
-import type { FindOptions } from '../../resource/base.ts'
-import { fnmatch } from '../../util/fnmatch.ts'
-import { FileType, PathSpec, type FileStat } from '../../types.ts'
-import { rstripSlash } from '../../util/slash.ts'
-
-export interface WalkFindDeps {
-  readdir: (spec: PathSpec, index?: IndexCacheStore) => Promise<string[]>
-  stat: (spec: PathSpec, index?: IndexCacheStore) => Promise<FileStat>
-  isDirName: (child: string) => boolean | null
-}
-
-interface WalkEntry {
-  path: string
-  depth: number
-  file: boolean
-}
-
-function isEnoent(err: unknown): boolean {
-  return err instanceof Error && (err as Error & { code?: string }).code === 'ENOENT'
-}
-
-function modifiedTs(modified: string | null): number | null {
-  // Naive timestamps are UTC, mirroring the Python implementation.
-  if (modified === null || modified === '') return null
-  const hasTz = /(?:Z|[+-]\d{2}:?\d{2})$/.test(modified)
-  const iso = hasTz ? modified : modified.includes(':') ? `${modified}Z` : `${modified}T00:00:00Z`
-  const ts = Date.parse(iso) / 1000
-  return Number.isNaN(ts) ? null : ts
-}
-
-async function statEntry(
-  deps: WalkFindDeps,
-  path: string,
-  prefix: string,
-  index: IndexCacheStore | undefined,
-): Promise<FileStat | null> {
-  const spec = new PathSpec({ original: path, directory: path, resolved: false, prefix })
-  try {
-    return await deps.stat(spec, index)
-  } catch (err) {
-    // Only missing entries resolve to null; API errors (rate limit, auth) propagate.
-    if (isEnoent(err)) return null
-    throw err
-  }
-}
-
-async function walk(
-  deps: WalkFindDeps,
-  spec: PathSpec,
-  index: IndexCacheStore | undefined,
-  maxDepth: number | null,
-  depth: number,
-  out: WalkEntry[],
-): Promise<void> {
-  if (maxDepth !== null && depth > maxDepth) return
-  let children: string[]
-  try {
-    children = await deps.readdir(spec, index)
-  } catch (err) {
-    // Only vanished dirs are skipped; API errors (rate limit, auth) propagate.
-    if (isEnoent(err)) return
-    throw err
-  }
-  for (const child of children) {
-    const hint = deps.isDirName(child)
-    const trimmed = child.endsWith('/') ? rstripSlash(child) : child
-    let isFolder: boolean
-    if (hint === null) {
-      // Warm index-cache entries carry no trailing slash, so fall back to stat.
-      const s = await statEntry(deps, trimmed, spec.prefix, index)
-      isFolder = s !== null && s.type === FileType.DIRECTORY
-    } else {
-      isFolder = hint
-    }
-    out.push({ path: trimmed, depth, file: !isFolder })
-    if (isFolder) {
-      const childSpec = new PathSpec({
-        original: trimmed,
-        directory: trimmed,
-        resolved: false,
-        prefix: spec.prefix,
-      })
-      await walk(deps, childSpec, index, maxDepth, depth + 1, out)
-    }
-  }
-}
-
-export async function walkFind(
-  path: PathSpec,
-  deps: WalkFindDeps,
-  options: FindOptions = {},
-  index?: IndexCacheStore,
-): Promise<string[]> {
-  const collected: WalkEntry[] = []
-  // GNU depth convention: the search root is depth 0, its children are
-  // depth 1, so the walk starts at 1 and -maxdepth 0 lists nothing.
-  await walk(deps, path, index, options.maxDepth ?? null, 1, collected)
-  const prefix = path.prefix
-  const results: string[] = []
-  // Codepoint comparison matches Python's sorted(); localeCompare does not.
-  collected.sort((a, b) => (a.path < b.path ? -1 : a.path > b.path ? 1 : 0))
-  for (const entry of collected) {
-    const name = entry.path.split('/').pop() ?? ''
-    if (options.minDepth != null && entry.depth < options.minDepth) continue
-    if (options.type === 'f' && !entry.file) continue
-    if (options.type === 'd' && entry.file) continue
-    if (options.orNames != null && options.orNames.length > 0) {
-      if (!options.orNames.some((pat) => fnmatch(name, pat))) continue
-    } else if (options.name != null && !fnmatch(name, options.name)) {
-      continue
-    }
-    if (options.iname != null && !fnmatch(name.toLowerCase(), options.iname.toLowerCase())) {
-      continue
-    }
-    const key =
-      prefix !== '' && entry.path.startsWith(prefix) ? entry.path.slice(prefix.length) : entry.path
-    if (options.pathPattern != null && !fnmatch(key, options.pathPattern)) continue
-    if (options.nameExclude != null && fnmatch(name, options.nameExclude)) continue
-    const needSize = entry.file && (options.minSize != null || options.maxSize != null)
-    const needMtime = options.mtimeMin != null || options.mtimeMax != null
-    if (needSize || needMtime) {
-      const st = await statEntry(deps, entry.path, prefix, index)
-      if (st === null) continue
-      if (needSize) {
-        const size = st.size ?? 0
-        if (options.minSize != null && size < options.minSize) continue
-        if (options.maxSize != null && size > options.maxSize) continue
-      }
-      if (needMtime) {
-        const mt = modifiedTs(st.modified)
-        if (mt === null) continue
-        if (options.mtimeMin != null && mt < options.mtimeMin) continue
-        if (options.mtimeMax != null && mt > options.mtimeMax) continue
-      }
-    }
-    results.push(key)
-  }
-  return results
-}
